hw/arm/apple-m1: Enough initial code to fully boot m1n1
The results of this patch is a build that will boot a stock compile
of m1n1 using the .elf version. You can redirect the serial to a pty
and then feed that pty to the m1n1 python code and get an interface.

M1N1 needs a small patch to properly handle a device tree that doesn't
contain the nodes it's looking for (checking validity instead of ignoring)
because I don't define a real device tree, just a stub to avoid crashes.

Real hardware would have a real tree which should always have these nodes.