hw/arm/apple-m1: Add ram framebuffer
This doesn't add any actual way to display it but that shouldn't be hard to
model next based on the code from ramfb. We can't use ramfb directly it seems
because it seems to use fw_cfg which is a structure I don't entirely
understand yet, I should do that and if possible support using the existing code.

It's likely that custom FB code for the M1 will be added anyway since the
M1 framebuffer (at least after boot) uses a 30 bit color mode which matches
likely nothing and there seems to be no hooks to support color conversion in
existing code.

This commit also removes the horrible hack from include/hw/elf_ops.h that allowed
m1n1 to load at the correct address. Because it's able to be loaded flat you can use
the generic loader support (-device loader,file=m1n1.elf(or macho),addr=0x10000) to
load it at the correct address and (-device loader,addr=0x14800,cpu-num=0) to boot it.

The first line sets it up to load into memory, the second command sets up the CPU to
change it's PC to 0x148000 which is the entry point _start (at least currently) with
the added offset. The boot_args defined in the machine code assume it's loaded at 0x10000
so you need to use that address for now.