hw/char/exynos4210_uart: Fix FIFO implementation
The Exynos4210 UART code has a duplicate FIFO implementation like many
other UART frontends. This wouldn't be a huge issue if it actually
worked properly. Unfortunately if you try to use it, once the FIFO is
full it will wrap around and think that it is empty again, causing the
data to be immediately written over by the next input set of data.

To fix this issue, move the Exynos4210 UART over to use a Fifo8, the
common qemu FIFO implementation, removing duplicated and broken code.