hw/arm/apple-m1: Miscellaneous dev pokes
Just a few bits and bobs for development, increase the UART buffer by a
lot to help with loading large amounts of data (it works well at 16
bytes but will work very well at a larger size since we almost always do
big writes from the host and little writes from the guest).

Also the AIC is stubbed in since Linux tries to use it.