hw/arm/apple-m1: Include realishy ADT
This replaces the stub 0 entry ADT with a small built-in one to specify
settings that m1n1 wants to see (and uses to configure the fdt it
creates to boot linux). Unfortunately creating an ADT dynamically is a
task in itself, so to avoid needing to patch m1n1 for testing like I've
been doing, or fully creating some sort of small ADT library, this code
just embeds the output of the valid tree converted to ADT format.