hw/arm/apple-m1: Initial cleanups
The current emulated M1 silicon support was very rough as it was the
product of just getting things to work. This is the start of the
eventual cleanup that will lead to upstream friendly code.

In short we:
* Define actual CPU core types for the Firestorm and Icestorm cores,
  along with giving them more registers (the Icestorm ones are still
  potentially missing their model specific registers since the main
  core is identifying as a firestorm and SMP bring up hasn't been
  completed.
* Moved M1 support from apple-m1-soc.c to apple-m1.c, this might end up
  being reverted, but the M1 is the "friendly" name of the SoC and we
  don't need an extraneous "SoC" added to the names
* Cleaned up the memory address handling a little so that the defines
  which aren't needed outside the .c file are setup only in that .c
  This prepares us for cleaning up the device tree and boot args
  handling to be based on putting them in RAM that's initialized on
  reset.

The names of the cores "aapl-firestorm" and "aapl-icestorm" are based on
the device tree names that upstream Linux will use, but they should
probably be renamed "apple-firestorm" and "apple-icestorm" since the
abbreviation isn't altogether useful.

There is still a lot of cleanup left, but more functionality is required
before that really makes sense.