hw/intc/apple-aic: Initial support for the Apple AIC
The AIC is the M1's interrupt controller, it's a fairly simple device
and this patch has initial support.

It is only about good enough to stub the device more properly, and in
fact it's less obvious what's going on
since there's no traces and it's not giving printouts
for unimplemented functionality now that everything is
properly stubbed. But this is a staging commit before
more proper support drops.