AIC: more work
Lots of work on getting the AIC to actually do something.

There's also some hacks in here that are supposed to get the UART to be
happy with the APPLE config flags but
the whole thing just doesn't work. I see irqs get raised
but linux just stalls like normal.

It's very confusing and will probably take awhile to debug.

In the process I realized the whole structure of the AIC driver is going
to be flawed since effectively each CPU
has a copy of several of the registers that's aliased to
the same memory region for ease of use. The current
code will need to change significantly to support that.

Furthermore a lot of the code is very low performance and
poor about the way it handles things. The only things
basically almost right are the IPIs which are not useful
since SMP isn't setup yet in QEMU.