hw/intc/apple-aic: Port to new bit functions
The initial attempts with using the standard QEMU bitmaps
was unsuccessful because they're rather coupled to their
definition as longs, and the AIC allows accessing the
registers at a native level, since this matches using
32-bit chunks for the bitmap, we locally implement
a bitmap that uses uint32_t chunks, allowing the
set/clear methods to natively set/clear chunks of the
bitmap.

This still has the massive hack from before (only supporting the single
serial IRQ) but with the new getters
it should be possible to make something much more reasonable.
Unfortunately the (optimized) find_first_bit/findn_next_bit aren't
applicable to our new uint32_t bitmap and we cannot copy the code
directly since these contributions are hopefully going to be MIT
licensed for now.