hw/intc/apple-aic: Functional state
This is the first version of AIC emulation that is at least
semi-functional. It doesn't properly account for multiple processors,
but with a single core system it should match enough hardware behavior
to boot Linux.

There's more registers (the per-cpu offsets specifically) that we should
implement but those will have to come after
proper support for SMP is implemented, since that really
needs to come next.