Formatting fixes
Also reintroduce a larger buffer, this doesn't fix the fact that M1N1
continuously polling on the buffer will cause the larger size to not
matter until the FIFO is used properly with interrupts.

Fixing that is probably better solved by adding support in M1N1 for the
FIFO properly (justification is it will speed up hardware loads during
testing), for QEMU we can fix the machine setup to use the -dtb, -initrd
and -kernel options to preload data into memory that we can jump into
without loading by passing their memory addresses into proxy function
calls.

"fat binaries" with the kernel appended to the end of the M1N1 image are
now supported too (which might be easier to use), and support for that
might make it possible to jump directly into the kernel loader which is
what we need for QEMU to seemlessly boot. (Never mind the fact that if
we
just use the standard ARM kernel loading code we can just start the
kernel properly and fill in the dtb ourselves like m1n1 would have...)