hw/arm/apple-m1: Add support for kernel options
This adds an initial support for using the -kernel, -initrd and -dtb
options to preload the M1 compatible kernel, dtb and initrd into memory
to later be filled by m1n1.

You still need to use the proxy to edit the
DTB and uncompress the kernel (since qemu is just loading it as a blob
into the guest's ram).

This means you still need to point -firmware to M1N1 and this is only
useful with a copy of a QEMU specific loading script for M1N1. Once I
add support for the latest M1N1 changes that support a concatenated
image QEMU should theoretically be able to immediately load into the
kernel via M1N1.