hw/arm/apple-m1: Initial support for SMP
This adds enough initial support to
manually start other cores along with
using a hacked M1N1 to do M1N1s SMP
startup (which creates a spintable
for the cores allowing M1N1 to run
SMP code). This will be used by Linux
to bring up the cores for Linux SMP
but this has not been tested yet.