hw/intc/apple-aic.c: Implement weird SMP support
This is like... the hackiest thing ever and certainly will have issue
if the MemoryRegionOps code is called from more than one context ever.
I'm fairly certain it's always called with an IO lock by default which
means that we probably need basic SMP barriers to make sure reads and
writes get posted and are fresh, but that's about it. Anyway there's no
point making broken code work. This is just something to see if this is
the issue with linux right now.

It's not enough on it's own still, but it's probably required to fix
whatever's the real problem. It was also written very quickly so almost
certainly I've missed something obvious.