Apple M1 code changes
This version (once bug fixed) will have "proper" support for SMP
with the AIC device. The only issue is that it's implemented using
the current_cpu variable which is a hack. Fixing this will make
things a lot messier at the moment and get in the way of debugging
the AIC changes so for now this is left for future improvement.