AIC fixes (first working SMP version)
This is the first version of the AIC code that actually functions
properly with SMP. It boots Linux correctly and IPIs are triggered
correctly.

It retains the current_cpu hack as that's the cleanest method to
solve the per-core mirroring issue at the present moment. In the
future a better solution will be devised. For now this is good enough.

This also changes the order of core initialization so the QEMU CPU
indexes match the AIC's concept of CPU indexes. This is something that
in the future will be resolved by having the device trees fed in be
mapped to the indexes used and specifically assigning CPU indices to
the cores used in the AIC.