hw/display/m1-fb: Cleanups
The FB code was pretty quickly written and pretty much just designed
to work. It makes operations such as trying to specify a size difficult.
The current code would still be hard to use to set a size since the
properties don't end up exposed as far as I know, but it's enough.

This also adds some VMState stuff for migration support, the AIC should
have this added next, but it shouldn't be considered stable until the FB
data structure is fleshed out (we need to know more about how it works
as an FB device, and probably the FB device will end up becoming the
GPU device with an FB component underneath as more begins to be
understood.

I can't decide what anything should be called and the code is
inconsitent in general about underscore seperators vs. hyphens. For now
I think underscores seem to be the path forward so I've converted this
code to use those and removed the "APPLE" prefix from the FB code since
M1 is enough to distinguish it fully. The AIC retains the Apple prefix
because the AIC isn't an "M1" component so much as it is
"AppleInterruptController" which the Apple prefix helps make clear.