M1 support: rework memory regions
Instead of putting the firmware payload, the ADT, and the boot args in ROM
space, they are put in RAM. This is closer to the way a real M1 device works.

The video RAM is also put at the end of the RAM space.

The firmware/ADT/boot-args RAM regions are implemented as “reloadable”; from
the guest point of view, they are modifiable, but when the machine is reset,
their content is reset to what the M1 bootloader would have put there
initially.