hw/arm/apple-m1: Fix frambuffer
Previous code tried to use a memory region alias as the framebuffer memory region.
This would work in theory but
seems to be incompatible somewhat with the current form
of the framebuffer helper code. Since there's no other way
to easily "slice up" a region into a carved out memory region
besides aliases which don't work here, we switch back to a
seperate allocation which is mapped into the main memory
to give the system bus access to it.

This change is tested and allows successfully booting M1N1 with framebuffer
graphics as long as M1N1 is patched to avoid GXF and SPRR functionality which
is not supported or stubbed out yet.