[hw/arm/apple-m1]: Fix framebuffer
The initial patch tried to use an alias of main RAM, this should work and
could if we used different code to handle the framebuffer (ramfb's code
probably works just fine), but we use the helper functions which have
trouble getting a MemoryRegionSection for the aliased region and therefore
break completely.

We revert that change to just overlay a vram specific section for now.